% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_risk.R
\name{br_group}
\alias{br_group}
\title{Posterior Samples for a Benefit/Risk Group}
\usage{
br_group(label, ...)
}
\arguments{
\item{label}{a string indicating the name of the group.}

\item{...}{named arguments which correspond to the names of the
benefits/risks specified by \code{benefit()} and \code{risk()} in a call to \code{br()}.}
}
\value{
A named list with the posterior samples and an assigned S3 class.
}
\description{
Posterior Samples for a Benefit/Risk Group
}
\details{
This function is intended to be used as an input argument to
the \code{br()} function.
}
\examples{
br(
  benefit("CV", function(x) x, weight = 0.5),
  risk("DVT", function(x) 2 * x, weight = 0.5),
  br_group(
    label = "PBO",
    samples("CV", rnorm(100)),
    samples("DVT", rnorm(100))
  ),
  br_group(
    label = "TRT",
    samples("CV", rnorm(100)),
    samples("DVT", rnorm(100))
  )
)

}
