% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.brisk_br}
\alias{plot.brisk_br}
\title{Plot Benefit/Risk Posterior Scores}
\usage{
\method{plot}{brisk_br}(x, reference = NULL, ...)
}
\arguments{
\item{x}{output from a call to \code{br()} or \code{mcda()}.}

\item{reference}{a string indicating which group is the reference group which
is used to subtract scores from other groups.}

\item{...}{additional arguments throw an error.}
}
\value{
A ggplot object plotting the posterior densities of the weighted
utility scores.
}
\description{
Plot Benefit/Risk Posterior Scores
}
