% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{sim_weights}
\alias{sim_weights}
\title{Simulate Normalized Weights}
\usage{
sim_weights(n, ...)
}
\arguments{
\item{n}{number of weights to simulate.}

\item{...}{vectors of length 2 indicating the lower
and upper bound (respectively) of the un-normalized weights.  At least one
set of bounds must be equal to each other (e.g. c(1, 1)) and be the largest
set of bounds in the set specified.}
}
\description{
Simulate Normalized Weights
}
\details{
The weights are normalized relative to a set of bounds which
are equal to each other (e.g. c(1, 1)), and also are the largest set of
bounds in the set specified.  See Example.
}
\examples{
w <- sim_weights(10, a = c(1, 1), b = c(.4, .6), c = c(.2, .3))
# ratio of b to a is between c(.4, .6) / c(1, 1)
summary(w$b / w$a)
# ratio of c to a is between c(.2, .3) / c(1, 1)
summary(w$c / w$a)

}
